Fixing movie model to adapt to the api response

+++ app/src/main/java/com/fibelatti/moviedb/models/Movie.java
@@ -60,7 +60,7 @@ public class Movie
     private String releaseDate;
     @SerializedName("revenue")
     @Expose
-    private Integer revenue;
+    private Double revenue;
     @SerializedName("runtime")
     @Expose
     private Integer runtime;
@@ -225,11 +225,11 @@ public class Movie
         this.releaseDate = releaseDate;
     }
 
-    public Integer getRevenue() {
+    public Double getRevenue() {
         return revenue;
     }
 
-    public void setRevenue(Integer revenue) {
+    public void setRevenue(Double revenue) {
         this.revenue = revenue;
     }
